Intermediate commit

+++ src/client/client_status.h
@@ -123,7 +123,7 @@ struct client_status_s {
     /**
      * Counter used for crash simulation feature
      */
-    int                             crash_count;
+    long                            crash_count;
 #endif
 };
 
@@ -401,7 +401,7 @@ extern "C" {
      * @return a writable reference to the number of times the crash point
      * was traversed
      */
-    static inline int *client_status_get_crash_count(client_status_t *cs) {
+    static inline long *client_status_get_crash_count(client_status_t *cs) {
         return &cs->crash_count;
     }
 #endif /* _CRASH */
